Cleanup warnings in the main project and SparCraft

+++ UAlbertaBot/Source/BaseLocationManager.cpp
@@ -471,7 +471,7 @@ BWAPI::TilePosition BaseLocationManager::getNextExpansion(BWAPI::Player player) 
         if (!closestBase || distanceFromHome < minDistance)
         {
             closestBase = base;
-            minDistance = distanceFromHome;
+            minDistance = (int)distanceFromHome;
         }
     }
