+++ UAlbertaBot/Source/CombatCommander.h
@@ -27,7 +27,7 @@ class CombatCommander
     void            updateScoutDefenseSquad(int currentFrame);
 	void            updateDefenseSquads(int currentFrame);
 	void            updateAttackSquads();
-	void updateAttackSquad(UAlbertaBot::Squad & mainAttackSquad, int maxUnitsInSquad);
+	void updateAttackSquad(UAlbertaBot::Squad & mainAttackSquad, size_t maxUnitsInSquad);
     void            updateDropSquads();
 	void            updateIdleSquad();
 	bool            isSquadUpdateFrame(int currentFrame) const;
